# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Service General including Multi-DB, BPJS, and SatuSehat integrations.",
        "title": "GoPrint Service General API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/satusehat/reference/patient": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari data pasien FHIR Satu Sehat berdasarkan kombinasi parameter (Nama, NIK, NIK Ibu, Tanggal Lahir, Gender)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Patient"
                ],
                "summary": "Pencarian Pasien (Satu Sehat) Multi-Parameter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nomor Induk Kependudukan",
                        "name": "nik",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nomor Induk Kependudukan Ibu (Untuk bayi)",
                        "name": "nik_ibu",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nama Pasien",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tanggal Lahir (YYYY-MM-DD)",
                        "name": "birthdate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Jenis Kelamin (male/female)",
                        "name": "gender",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendaftarkan data pasien baru ke API Satu Sehat dan mengembalikan IHS Number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Patient"
                ],
                "summary": "Daftar Pasien Baru (Satu Sehat)",
                "parameters": [
                    {
                        "description": "Data Pasien Baru",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_internal_satusehat_reference_patient.CreatePatientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/patient/nik/{nik}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari data pasien FHIR Satu Sehat berdasarkan Nomor Induk Kependudukan (NIK)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Patient"
                ],
                "summary": "Cari Pasien (Satu Sehat) berdasarkan NIK",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nomor Induk Kependudukan",
                        "name": "nik",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/patient/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari data pasien FHIR Satu Sehat berdasarkan IHS Number / ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Patient"
                ],
                "summary": "Cari Pasien (Satu Sehat) berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IHS Number / Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/practitioner": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari data Practitioner FHIR Satu Sehat berdasarkan parameter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Practitioner"
                ],
                "summary": "Pencarian Tenaga Medis (Satu Sehat) Multi-Parameter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nomor Induk Kependudukan",
                        "name": "nik",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nama Tenaga Medis",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Jenis Kelamin (male/female)",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tanggal Lahir (YYYY-MM-DD)",
                        "name": "birthdate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/practitioner/nik/{nik}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari data Practitioner FHIR Satu Sehat berdasarkan NIK",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Practitioner"
                ],
                "summary": "Cari Tenaga Medis (Satu Sehat) berdasarkan NIK",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nomor Induk Kependudukan",
                        "name": "nik",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/practitioner/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari data Practitioner FHIR Satu Sehat berdasarkan IHS Number / ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Practitioner"
                ],
                "summary": "Cari Tenaga Medis (Satu Sehat) berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IHS Number / Practitioner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "service_internal_satusehat_reference_patient.CreatePatientRequest": {
            "type": "object",
            "required": [
                "birth_date",
                "gender",
                "name",
                "nik"
            ],
            "properties": {
                "address": {
                    "description": "Opsional",
                    "type": "string"
                },
                "birth_date": {
                    "description": "Format: YYYY-MM-DD",
                    "type": "string"
                },
                "gender": {
                    "description": "male, female, other, unknown",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "phone": {
                    "description": "Opsional",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}